Rodrigo Grazini - J-B-18-Rodri-Grazi - Added lombok support, removed unnecesary comments and this pointers, replaced Connection implementations with the singleton implementations
#If applied this commit will...
* Add lombok support
* Implement the connections as singletons, replacing the ${ConnectionType}ConnectionSingletons classes
* remove redundant comments and "this" pointers

#why?
* Lombok for less typing :)
* The sinlgetons were corrected regarding [this subject] (https://github.com/Wasini/Java-bootcamp-2018-Goal0-Rodrigo-Grazini/pull/1#discussion_r219489481)
* Some comments were unnecesary
