+++ pom.xml
@@ -27,6 +27,12 @@
         <artifactId>log4j</artifactId>
         <version>1.2.17</version>
     </dependency>
+    <dependency>
+      <groupId>org.projectlombok</groupId>
+      <artifactId>lombok</artifactId>
+      <version>1.16.22</version>
+    <scope>provided</scope>
+    </dependency>
   </dependencies>
 
 </project>
