+++ src/main/java/com/globant/bootcamp/App.java
@@ -1,14 +1,10 @@
 package com.globant.bootcamp;
 
-import com.globant.bootcamp.patterns.sinlgeton.SqlConnectionSingleton;
+import com.globant.bootcamp.patterns.sinlgeton.SqlConnection;
 import com.globant.bootcamp.repository.Connection;
 
 import java.util.Properties;
 
-/**
- * Hello world!
- *
- */
 public class App {
 
 	public static void main(String[] argsStrings) {
@@ -16,7 +12,7 @@ public class App {
 		credentials.put("user", "admin");
 		credentials.put("password", "admin");
 		
-		Connection connection = SqlConnectionSingleton.getInstance("store");
+		Connection connection = SqlConnection.getInstance("store");
 		connection.connect(credentials);
 	}
 }
