--- src/main/java/com/globant/bootcamp/patterns/sinlgeton/PostgresConnectionSingleton.java
@@ -1,23 +1,0 @@
-package com.globant.bootcamp.patterns.sinlgeton;
-
-import org.apache.log4j.Logger;
-import com.globant.bootcamp.repository.Connection;
-import com.globant.bootcamp.repository.PostgresConnection;
-
-/**
- * Singleton of a database connection
- */
-public class PostgresConnectionSingleton {
-
-	final static Logger logger = Logger.getLogger(PostgresConnectionSingleton.class);
-
-	private static Connection instance;
-
-	public static Connection getInstance(String dbName) {
-		if (instance == null) {
-			logger.info("Creating static instance of Postgres connection");
-			instance = new PostgresConnection(dbName);
-		}
-		return instance;
-	}
-}
