+++ src/main/java/com/globant/bootcamp/patterns/sinlgeton/SqlConnection.java
@@ -1,26 +1,40 @@
-package com.globant.bootcamp.repository;
-
-import org.apache.log4j.Logger;
+package com.globant.bootcamp.patterns.sinlgeton;
 
 import java.util.Properties;
 import java.util.ResourceBundle;
+
+import lombok.Getter;
+
+import org.apache.log4j.Logger;
+
+import com.globant.bootcamp.repository.Connection;
 
 /**
  * Stub class for a "sql database connection" with static user and password authentication
  */
 public class SqlConnection implements Connection {
 
+	private static Connection instance;
+
     final static Logger logger = Logger.getLogger(SqlConnection.class);
 
-	private String dbUrl;
+	@Getter private String Url;
 
-	private boolean status;
+	@Getter private boolean open;
 
-	public SqlConnection(String dbName) {
+	private SqlConnection(String dbName) {
 		if (dbName == null) {
 			throw new IllegalArgumentException("Database URL can't be null");
 		}
-		this.dbUrl = "mysql:postgres:" + dbName;
+		this.Url = "mysql:postgres:" + dbName;
+	}
+
+	public static Connection getInstance(String dbName) {
+		if (instance == null) {
+			logger.info("Creating static instance of Sql connection");
+			instance = new SqlConnection(dbName);
+		}
+		return instance;
 	}
 	
 	@Override
@@ -28,27 +42,18 @@ public class SqlConnection implements Connection {
 	    ResourceBundle labels = ResourceBundle.getBundle("credentials");
 	    String user = labels.getString("mysql.user");
 	    String password = labels.getString("mysql.password");
-	    logger.info("Using stub mysql connection class\n Connecting to mysql server with adress; "+this.getUrl());
+	    logger.info("Using stub mysql connection class\n Connecting to mysql server with adress; " + getUrl());
 	    try{
             Thread.sleep(1000);
         } catch(InterruptedException ex) {
             Thread.currentThread().interrupt();
         }
-	    this.status = user.equals(credentials.get("user")) && password.equals(credentials.get("password"));   
-	    if (this.getStatus()){
+	    open = user.equals(credentials.get("user")) && password.equals(credentials.get("password"));   
+	    if (isOpen()){
 	        logger.info("User "+credentials.get("user")+ " connected successfully");
 	    } else {
 	        logger.error("Something went wrong while connecting, incorrect user or password?");
 	    }
     }
-    
-    @Override
-    public boolean getStatus() {
-        return this.status;
-    }
-    
-    @Override
-    public String getUrl() {
-        return this.dbUrl;
-    }
 }
+
