+++ src/main/java/com/globant/bootcamp/repository/Connection.java
@@ -2,19 +2,13 @@ package com.globant.bootcamp.repository;
 
 import java.util.Properties;
 
-/**
- * Contract for the Connection class
- */
 public interface Connection {
 	/**
 	 * Tries to connect to a database using the given properties with keys for "user" and "password"
 	 */
-	public abstract void connect(Properties credentials);
+	void connect(Properties credentials);
+	
+	boolean isOpen();
 
-	/**
-	 * @return status: If connection with the database is established 
-	 */
-	public boolean getStatus();
-
-	public String getUrl();
+	String getUrl();
 }
